Fixed weird compiler & linker warnings on OS X
git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc/trunk@3010 0db28fee-5c31-0410-96ea-e69cce70e049

+++ engine/src/rgbmatrix.cpp
@@ -278,7 +278,7 @@ void RGBMatrix::tap()
     {
         FixtureGroup* grp = doc()->fixtureGroup(fixtureGroup());
         // Filter out taps that are too close to each other
-        if (grp != NULL && m_roundTime->elapsed() >= (duration() / 4))
+        if (grp != NULL && uint(m_roundTime->elapsed()) >= (duration() / 4))
             roundCheck(grp->size());
     }
 }
